Refactor PosgreSQL API
