+++ src/main.py
@@ -6,11 +6,11 @@ from concurrent.futures import ProcessPoolExecutor
 from fastapi import FastAPI
 
 # 引入 pg api router
-from .db.pg_api import router as pg_router
+from .app.route import router
 
 app = FastAPI()
 
-app.include_router(pg_router)
+app.include_router(router)
 
 
 @app.get("/")
